callbacks: move callback helpers to another file
We have callbacks for error handling and we will introduce more
non-event callbacks in the future. Just move the appropriate bit into a
separate file and document how to build a new callback.
